move JSON parser to zsio
